@@ -1,6 +1,9 @@
 <?php namespace Zephyrus\Application;
 
 use InvalidArgumentException;
+use ReflectionClass;
+use ReflectionMethod;
+use RuntimeException;
 use Zephyrus\Exceptions\RouteArgumentException;
 use Zephyrus\Network\ContentType;
 use Zephyrus\Network\Request;
@@ -11,13 +14,17 @@ use Zephyrus\Network\Responses\RenderResponses;
 use Zephyrus\Network\Responses\StreamResponses;
 use Zephyrus\Network\Responses\SuccessResponse;
 use Zephyrus\Network\Responses\XmlResponses;
-use Zephyrus\Network\Routable;
-use Zephyrus\Network\Router;
+use Zephyrus\Network\Router\Delete;
+use Zephyrus\Network\Router\Get;
+use Zephyrus\Network\Router\Patch;
+use Zephyrus\Network\Router\Post;
+use Zephyrus\Network\Router\Put;
+use Zephyrus\Network\RouteRepository;
 
-abstract class Controller implements Routable
+abstract class Controller
 {
     protected ?Request $request = null;
-    private Router $router;
+    private ?RouteRepository $repository = null;
     private array $overrideArguments = [];
     private array $restrictedArguments = [];
 
@@ -28,10 +35,61 @@ abstract class Controller implements Routable
     use XmlResponses;
     use DownloadResponses;
 
-    public function __construct(Router $router)
+    /**
+     * Defines all the routes supported by this controller associated with inner methods.
+     */
+    public function initializeRoutes(): void
     {
-        $this->router = $router;
-        $this->request = &$router->getRequest();
+    }
+
+    public function initializeRoutesFromAttributes(RouteRepository $repository): void
+    {
+        $reflection = new ReflectionClass($this);
+        $methods = $reflection->getMethods(ReflectionMethod::IS_PUBLIC);
+        $supportedAttributes = [Get::class, Post::class, Put::class, Patch::class, Delete::class];
+        foreach ($methods as $method) {
+            $attributes = $method->getAttributes();
+            foreach ($attributes as $attribute) {
+                if (in_array($attribute->getName(), $supportedAttributes)) {
+                    $instance = $attribute->newInstance();
+                }
+                switch ($attribute->getName()) {
+                    case Get::class:
+                        $repository->get($instance->getRoute(), [$this, $method->name]);
+                        break;
+                    case Post::class:
+                        $repository->post($instance->getRoute(), [$this, $method->name]);
+                        break;
+                    case Put::class:
+                        $repository->put($instance->getRoute(), [$this, $method->name]);
+                        break;
+                    case Patch::class:
+                        $repository->patch($instance->getRoute(), [$this, $method->name]);
+                        break;
+                    case Delete::class:
+                        $repository->delete($instance->getRoute(), [$this, $method->name]);
+                        break;
+                }
+            }
+        }
+    }
+
+    /**
+     * Applies the route collection instance to be used with the inner get, post, put, patch and delete method. These
+     * will register the route into the given repository. This method is necessary to make sure we keep the controller
+     * instance reference.
+     *
+     * @param RouteRepository $repository
+     * @return void
+     */
+    public function setRouteRepository(RouteRepository $repository): void
+    {
+        $this->repository = $repository;
+    }
+
+    public function setRequest(Request $request): void
+    {
+        $this->request = &$request;
     }
 
     /**
@@ -68,7 +126,7 @@ abstract class Controller implements Routable
      * @param string $argumentName
      * @param callable $callback
      */
-    public function overrideArgument(string $argumentName, callable $callback)
+    public function overrideArgument(string $argumentName, callable $callback): void
     {
         if (count((new Callback($callback))->getReflection()->getParameters()) != 1) {
             throw new InvalidArgumentException("Override callback should have only one argument which will contain the value of the associated argument name");
@@ -84,7 +142,7 @@ abstract class Controller implements Routable
      * @param string $parameterName
      * @param Rule[] $rules
      */
-    public function restrictArgument(string $parameterName, array $rules)
+    public function restrictArgument(string $parameterName, array $rules): void
     {
         foreach ($rules as $rule) {
             if (!($rule instanceof Rule)) {
@@ -139,9 +197,12 @@ abstract class Controller implements Routable
      * @param string $instanceMethod
      * @param string | array $acceptedFormats
      */
-    final protected function get(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY)
+    final protected function get(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY): void
     {
-        $this->router->get($uri, [$this, $instanceMethod], $acceptedFormats);
+        if (is_null($this->repository)) {
+            throw new RuntimeException("You must first set a RouteRepository instance on which the route definition will apply. Be sure to use the setRouteRepository method before any calls to get, post, put, patch and delete methods.");
+        }
+        $this->repository->get($uri, [$this, $instanceMethod], $acceptedFormats);
     }
 
     /**
@@ -154,9 +215,12 @@ abstract class Controller implements Routable
      * @param string $instanceMethod
      * @param string | array $acceptedFormats
      */
-    final protected function post(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY)
+    final protected function post(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY): void
     {
-        $this->router->post($uri, [$this, $instanceMethod], $acceptedFormats);
+        if (is_null($this->repository)) {
+            throw new RuntimeException("You must first set a RouteRepository instance on which the route definition will apply. Be sure to use the setRouteRepository method before any calls to get, post, put, patch and delete methods.");
+        }
+        $this->repository->post($uri, [$this, $instanceMethod], $acceptedFormats);
     }
 
     /**
@@ -169,9 +233,12 @@ abstract class Controller implements Routable
      * @param string $instanceMethod
      * @param string | array $acceptedFormats
      */
-    final protected function put(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY)
+    final protected function put(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY): void
     {
-        $this->router->put($uri, [$this, $instanceMethod], $acceptedFormats);
+        if (is_null($this->repository)) {
+            throw new RuntimeException("You must first set a RouteRepository instance on which the route definition will apply. Be sure to use the setRouteRepository method before any calls to get, post, put, patch and delete methods.");
+        }
+        $this->repository->put($uri, [$this, $instanceMethod], $acceptedFormats);
     }
 
     /**
@@ -185,9 +252,12 @@ abstract class Controller implements Routable
      * @param string $instanceMethod
      * @param string | array $acceptedFormats
      */
-    final protected function patch(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY)
+    final protected function patch(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY): void
     {
-        $this->router->patch($uri, [$this, $instanceMethod], $acceptedFormats);
+        if (is_null($this->repository)) {
+            throw new RuntimeException("You must first set a RouteRepository instance on which the route definition will apply. Be sure to use the setRouteRepository method before any calls to get, post, put, patch and delete methods.");
+        }
+        $this->repository->patch($uri, [$this, $instanceMethod], $acceptedFormats);
     }
 
     /**
@@ -201,9 +271,12 @@ abstract class Controller implements Routable
      * @param string $instanceMethod
      * @param string | array $acceptedFormats
      */
-    final protected function delete(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY)
+    final protected function delete(string $uri, string $instanceMethod, string|array $acceptedFormats = ContentType::ANY): void
     {
-        $this->router->delete($uri, [$this, $instanceMethod], $acceptedFormats);
+        if (is_null($this->repository)) {
+            throw new RuntimeException("You must first set a RouteRepository instance on which the route definition will apply. Be sure to use the setRouteRepository method before any calls to get, post, put, patch and delete methods.");
+        }
+        $this->repository->delete($uri, [$this, $instanceMethod], $acceptedFormats);
     }
 
     /**
