@@ -2,9 +2,10 @@
 
 use PHPUnit\Framework\TestCase;
 use Zephyrus\Application\Bootstrap;
-use Zephyrus\Network\Router;
 use Zephyrus\Network\ContentType;
 use Zephyrus\Network\Request;
+use Zephyrus\Network\Router;
+use Zephyrus\Network\RouteRepository;
 
 class BootstrapTest extends TestCase
 {
@@ -25,14 +26,21 @@ class BootstrapTest extends TestCase
         $server['SERVER_PORT'] = '80';
         $server['CONTENT_TYPE'] = ContentType::PLAIN;
 
+        $repository = new RouteRepository();
+        Bootstrap::initializeControllerRoutes($repository);
+
         // Mimics workflow
         $req = new Request('http://test.local/batman', 'get', [
             'server' => $server
         ]);
-        $router = new Router();
-        Bootstrap::initializeRoutableControllers($router);
-        ob_start();
-        $router->run($req);
-        self::assertEquals('batman rocks!', ob_get_clean());
+        $router = new Router($repository);
+        $response = $router->resolve($req);
+        self::assertEquals('batman rocks!', $response->getContent());
+
+        $req = new Request('http://test.local/robin', 'get', [
+            'server' => $server
+        ]);
+        $response = $router->resolve($req);
+        self::assertEquals('robin test', $response->getContent());
     }
 }
